마이닝 정책 개편 - 4

+++ database/migrations/2025_09_19_154055_create_level_matchings_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('bonus_id')->constrained('level_bonuses');
             $table->foreignId('transfer_id')->constrained('income_transfers');
-            $table->unsignedSmallInteger('referrer_id')->comment('산하 ID');
+            $table->unsignedBigInteger('referrer_id')->comment('산하 ID');
             $table->decimal('matching', 20, 9)->comment('레벨 매칭');
             $table->timestamps();
         });
